Placement restrictions (#487)
* Restrict under and above

Fixed crashing due to user trying to place blocks under the world, moving the above check to take effect, and making it send the player the error message.

* Changed TextComponent Translate

Due to a circular encoding problem changed "with" value type of Translate until TextComponent rework.

* Restrict on adventure and spectator

Added a check and error when user tries to place a block on adventure or spectator gamemodes.